# 🚀 Release v1.0.0: Hybrid Go/Python Architecture