… .golangci.yml
… ARCHITECTURE.md
… PROJECT_STRUCTURE.md
… README.md
… README_CLI.md
… RELEASE_CHECKLIST.md
… api/proto/webhook.pb.go
… api/proto/webhook_grpc.pb.go
… api/proto/webhook_pb2.py
… api/proto/webhook_pb2_grpc.py
… config.dev.yaml
… config.example.yaml
… config.prod.yaml
… config.yaml
… docker-compose.yml
… go.sum
… internal/cli/build.go
… internal/cli/commands.go
… internal/cli/service.go
… internal/cli/start.go
… internal/config/config.go
… internal/config/config_test.go
… internal/config/directories.go
… internal/config/manager.go
… internal/grpc/client.go
… internal/logging/manager.go
… internal/python/detector.go
… internal/python/manager.go
… internal/server/errors.go
… internal/server/logging.go